improved player control and web jitter fix.

+++ src/Player.hpp
@@ -44,7 +44,7 @@ private:
 
     int windowWidth{};
     int windowHeight{};
-    int scale{};
+    float scale{};
     int playerSize{};
 
     Rectangle playerRect{};
